feature[ats_ATS139673]: Consolidate Purge Transaction

+++ plugins/org.eclipse.osee.orcs.db/src/org/eclipse/osee/orcs/db/internal/callable/PurgeTransactionTxCallable.java
@@ -115,10 +115,9 @@ public class PurgeTransactionTxCallable extends AbstractDatastoreTxCallable<Inte
             txIdToDelete);
 
          //Find affected items
-         Map<Integer, IdJoinQuery> arts = findAffectedItems(connection, "art_id", "osee_artifact", txsToDelete);
-         Map<Integer, IdJoinQuery> attrs = findAffectedItems(connection, "attr_id", "osee_attribute", txsToDelete);
-         Map<Integer, IdJoinQuery> rels =
-            findAffectedItems(connection, "rel_link_id", "osee_relation_link", txsToDelete);
+         Map<Long, IdJoinQuery> arts = findAffectedItems(connection, "art_id", "osee_artifact", txsToDelete);
+         Map<Long, IdJoinQuery> attrs = findAffectedItems(connection, "attr_id", "osee_attribute", txsToDelete);
+         Map<Long, IdJoinQuery> rels = findAffectedItems(connection, "rel_link_id", "osee_relation_link", txsToDelete);
 
          //Update Baseline txs for Child Branches
          setChildBranchBaselineTxs(connection, txIdToDelete, previousTransactionId);
@@ -139,11 +138,11 @@ public class PurgeTransactionTxCallable extends AbstractDatastoreTxCallable<Inte
       return purgeCount;
    }
 
-   private void computeNewTxCurrents(JdbcConnection connection, Collection<Object[]> updateData, String itemId, String tableName, Map<Integer, IdJoinQuery> affected) throws OseeCoreException {
+   private void computeNewTxCurrents(JdbcConnection connection, Collection<Object[]> updateData, String itemId, String tableName, Map<Long, IdJoinQuery> affected) throws OseeCoreException {
       String query = String.format(FIND_NEW_TX_CURRENTS, tableName, itemId);
 
-      for (Entry<Integer, IdJoinQuery> entry : affected.entrySet()) {
-         Integer branchUuid = entry.getKey();
+      for (Entry<Long, IdJoinQuery> entry : affected.entrySet()) {
+         Long branchUuid = entry.getKey();
          IdJoinQuery joinQuery = entry.getValue();
          try {
             JdbcStatement statement = getJdbcClient().getStatement(connection);
@@ -173,13 +172,13 @@ public class PurgeTransactionTxCallable extends AbstractDatastoreTxCallable<Inte
       }
    }
 
-   private Map<Integer, IdJoinQuery> findAffectedItems(JdbcConnection connection, String itemId, String itemTable, List<Object[]> bindDataList) throws OseeCoreException {
-      Map<Integer, IdJoinQuery> items = new HashMap<Integer, IdJoinQuery>();
+   private Map<Long, IdJoinQuery> findAffectedItems(JdbcConnection connection, String itemId, String itemTable, List<Object[]> bindDataList) throws OseeCoreException {
+      Map<Long, IdJoinQuery> items = new HashMap<Long, IdJoinQuery>();
       JdbcStatement statement = getJdbcClient().getStatement(connection);
 
       try {
          for (Object[] bindData : bindDataList) {
-            Integer branchUuid = (Integer) bindData[0];
+            Long branchUuid = (Long) bindData[0];
             String query = String.format(SELECT_AFFECTED_ITEMS, itemId, itemTable);
             statement.runPreparedQuery(JdbcConstants.JDBC__MAX_FETCH_SIZE, query, bindData);
             IdJoinQuery joinId = joinFactory.createIdJoinQuery();
